Freeing memory in blocks.c; my incomplete attempt at i/o
Just added a function to free the dynamically allocated memory in
blocks.c

Added a file I started work on yesterday to do the i/o, without having
to specify a number of rows or columns (i.e. can give it any matrix
size input and it’ll work). Not *complete* yet, but I might toss it if
we’re only working with one data file