Started writing sequential collision detection
I am focussing on getting simple sequential code working before giving
tremendous thought to parallelisation. I have started some work on
collision detection, and have consequently created a new struct for
collisions. This is because a collision will involve multiple blocks,
and we need to know the rows of all of those blocks (for results). So I
decided a struct would be a good way to do that.

I have started writing a function to detect collisions, but there is a
segfault occurring in there, so we need to look into what’s causing
that. I have also written a method to free all memory allocated for
collision database (similar to methods for freeing memory allocated for
key databases and block databases)