Merged my work with main
Hopefully we can avoid these collisions in the future, since they’re
kind of annoying to deal with. Also, I changed allRows and allCols back
to ROWS and COLS because I think that’s much more readable.

Please do not delete the branch on sequential collision detection,
though. I’d like to leave it there in case we encounter issues in the
future.

The code should now successfully find all collisions in the matrix,
sequentially. We may have to modify the struct and code slightly to
print out exactly what is specified in the project prompt, but such
changes will be very trivial at this stage (just adding more fields to
the struct etc.).

The next (and most important) step is to parallelise. I’ve got a few
ideas as to how we could do this somewhat efficiently with little
modification to the existing code, but it’d be best to meet up and
discuss parallelisation first.