Wrote a simple parallelisation of block generation
Didn’t parallelise by column, since most blocks are in the last column.
Instead, I paralysed by the inner search. The biggest bottleneck in the
paralysation is the critical section at the bottom, where the partial
database and complete databases are merged together. We should focus on
trying to get rid of that completely (i.e. replace with atomic
somehow). I will now test the code for correctness and performance