Fixed tasking issue, and identified performance bottleneck
Identified correctness issue in tasking and fixed it; the parallel code
now runs correctly (i.e. returns same results as sequential code), but
has no performance improvement. Note that there is also no performance
loss now, which is miles ahead of where we were yesterday.

Also, I know exactly where the bottleneck in current program is, and
how to fix it: The last column is parallelised using tasks; a single
thread generates a task for each chunk of rows (i.e 0-1000, 1000-2000,
2000-3000), and other threads work on the tasks. The problem is that
the chunks of rows do not have an even amount of work in them: rows
0-1000 have a much, much larger number of blocks in them than any other
chunk. And rows 2000-4000 have comparatively few blocks. As such, some
threads are finishing much earlier than others, and there is a
performance bottleneck there. We will need to fix the way we divide the
work in the tasks, and I believe we will then get a reasonable speed up.