Added to notes
Added lots more to the notes on my approach to parallelisation.

Also, I believe the issue with the last column that is causing the
slowdown is a cache issue. Basically, the last column runs faster the
less threads there are executing at once. This suggests, to me, that
the threads are interfering with each other. This is most likely
something to do with cache (i.e. cache-miss ratio is skyrocketing when
multiple threads are running). This is probably the next thing to look
at.