Did some more testing; added to notes
Tested time spent in critical region to see if that was causing the
slow-down; it wasn’t, as very little time is being spent in there
(roughly 7 milli-seconds vs. 310 milli-seconds for overall execution).
Also tested the load-balancing, but found it to be quite even. Will
test for memory issues next.